Update type for contract item

+++ EVEStandard/Models/ContractItem.cs
@@ -39,7 +39,7 @@ namespace EVEStandard.Models
         /// </summary>
         /// <value>Unique ID for the item</value>
         [JsonProperty("record_id")]
-        public long RecordId { get; set; }
+        public int RecordId { get; set; }
 
         /// <summary>
         /// Type ID for item
